refactor(backend): extract codex thread scoping helpers
